Updated safe divider code

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
     try:
-        quotient = int(numerator) / int(denominator)
+        quotient = float(numerator) / float(denominator)
         print(f"The result of the division is {quotient}")
     except ValueError:
         print(f"Error: Please enter numeric values only.")
